+++ native/src/seal/encryptor.h
@@ -119,7 +119,7 @@ namespace seal
         /**
         Encrypts a plaintext with the public key and stores the result in
         destination.
-        
+
         The encryption parameters for the resulting ciphertext correspond to:
         1) in BFV, the highest (data) level in the modulus switching chain,
         2) in CKKS, the encryption parameters of the plaintext.
@@ -145,7 +145,7 @@ namespace seal
         /**
         Encrypts a zero plaintext with the public key and stores the result in
         destination.
-        
+
         The encryption parameters for the resulting ciphertext correspond to the
         highest (data) level in the modulus switching chain. Dynamic memory
         allocations in the process are allocated from the memory pool pointed to
@@ -166,7 +166,7 @@ namespace seal
         /**
         Encrypts a zero plaintext with the public key and stores the result in
         destination.
-        
+
         The encryption parameters for the resulting ciphertext correspond to the
         given parms_id. Dynamic memory allocations in the process are allocated
         from the memory pool pointed to by the given MemoryPoolHandle.
@@ -189,7 +189,7 @@ namespace seal
         /**
         Encrypts a plaintext with the secret key and stores the result in
         destination.
-        
+
         The encryption parameters for the resulting ciphertext correspond to:
         1) in BFV, the highest (data) level in the modulus switching chain,
         2) in CKKS, the encryption parameters of the plaintext.
@@ -263,7 +263,7 @@ namespace seal
         /**
         Encrypts a plaintext with the secret key and returns the ciphertext as
         a serializable object.
-        
+
         The encryption parameters for the resulting ciphertext correspond to:
         1) in BFV, the highest (data) level in the modulus switching chain,
         2) in CKKS, the encryption parameters of the plaintext.
@@ -324,7 +324,7 @@ namespace seal
         /**
         Encrypts a zero plaintext with the secret key and returns the ciphertext
         as a serializable object.
-        
+
         The encryption parameters for the resulting ciphertext correspond to the
         highest (data) level in the modulus switching chain. Dynamic memory
         allocations in the process are allocated from the memory pool pointed to